+++ tiger-common/src/test/java/de/gematik/test/tiger/common/config/TigerConfigurationTest.java
@@ -19,6 +19,7 @@ import java.io.File;
 import java.io.IOException;
 import java.net.ServerSocket;
 import java.nio.charset.StandardCharsets;
+import java.time.LocalDate;
 import java.util.List;
 import java.util.Map;
 import lombok.Builder;
@@ -432,6 +433,15 @@ public class TigerConfigurationTest {
                     Users.builder().username("guest").password("guest1234").roles(List.of("VIEW")).build());
     }
 
+    @Test
+    void shouldParseJava8Date() {
+        TigerGlobalConfiguration.reset();
+        TigerGlobalConfiguration.readFromYaml("users.blub: 2007-12-24T18:21Z");
+        final Users users = TigerGlobalConfiguration.instantiateConfigurationBean(Users.class, "users").get();
+        assertThat(users.blub)
+            .isEqualTo("2007-12-24");
+    }
+
     /**
      * ${ENV => GlobalConfigurationHelper.getString() ${json-unit.ignore} => interessiert dann folglich nicht
      * ${VAR.foobar} => GlobalConfigurationHelper.getSourceByName("VAR").getString()
@@ -712,6 +722,7 @@ public class TigerConfigurationTest {
     public static class Users {
         private String username;
         private String password;
+        private LocalDate blub;
         private List<String> roles;
     }
 }